Moves find_citations helper functions to helpers.py.
